model_mod.f90 and location_mod.f90 no longer have i8 indices where they are not needed.
The change to obs_def_navdas_mod.f90 is a simple line wrap.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/coamps@13205 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 829e5c9373ac218c272d7d8a8a8ebc29685a2832

+++ assimilation_code/location/threed_sphere/location_mod.f90
@@ -1385,7 +1385,7 @@ subroutine get_close_obs(gc, base_loc, base_type, locs, loc_qtys, loc_types, &
 type(get_close_type),          intent(in)  :: gc
 type(location_type),           intent(inout) :: base_loc, locs(:)
 integer,                       intent(in)  :: base_type, loc_qtys(:), loc_types(:)
-integer(i8),                   intent(out) :: num_close, close_ind(:)
+integer,                       intent(out) :: num_close, close_ind(:)
 real(r8),            optional, intent(out) :: dist(:)
 type(ensemble_type), optional, intent(in)  :: ens_handle
 
@@ -1407,7 +1407,7 @@ type(get_close_type),          intent(in)  :: gc
 type(location_type),           intent(inout)  :: base_loc, locs(:)
 integer,                       intent(in)  :: base_type, loc_qtys(:)
 integer(i8),                   intent(in)  :: loc_indx(:)
-integer(i8),                   intent(out) :: num_close, close_ind(:)
+integer,                       intent(out) :: num_close, close_ind(:)
 real(r8),            optional, intent(out) :: dist(:)
 type(ensemble_type), optional, intent(in)  :: ens_handle
 
@@ -1428,7 +1428,7 @@ subroutine get_close(gc, base_loc, base_type, locs, loc_qtys, &
 type(get_close_type),          intent(in)  :: gc
 type(location_type),           intent(inout)  :: base_loc, locs(:)
 integer,                       intent(in)  :: base_type, loc_qtys(:)
-integer(i8),                   intent(out) :: num_close, close_ind(:)
+integer,                       intent(out) :: num_close, close_ind(:)
 real(r8),            optional, intent(out) :: dist(:)
 type(ensemble_type), optional, intent(in)  :: ens_handle
